Screen Scrolling
Screen Scrolling is working with the buildings staying put. The fix
included creating a tiny cube to parent the building sprites to.

I also added a script to the building sprite so that we should be able
to randomly assign different building sprites to the object at runtime.

Saved what could be our initial level and moved all of the testing
scenes into a testing folder within the scenes folder.

Having issues using Leah's texture for the ground.